♻️ refactor(monitoring): extract codex probe model
Move auth-file normalization, quota-window decisions, and single-account probing into a dedicated model module.

Keep the session orchestrator focused on scheduling progress and result settlement.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
